hackathon/requirements.txt
# Forensica AI — Backend Dependencies
# ─────────────────────────────────────
# Install core deps:   pip install -r requirements.txt
# Install all deps:    pip install -r requirements.txt -r requirements-optional.txt

# ── Core (always required) ─────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
Pillow>=10.2.0
numpy>=1.26.0
pytesseract>=0.3.10
PyMuPDF>=1.24.0
requests>=2.32.0
firebase-admin>=6.5.0
python-dotenv>=1.0.1
opencv-python>=4.9.0
cryptography>=42.0.0

# ── Document processing (optional — backend degrades gracefully without these)
# Uncomment to enable real detection instead of intelligent stubs:

# PyMuPDF>=1.24.0            # PDF parsing (fitz) — font extraction + page count
# pytesseract>=0.3.10        # OCR wrapper (also needs Tesseract binary installed)
# Pillow>=10.2.0             # Image I/O for OCR + preprocessing
# opencv-python>=4.9.0       # Pixel anomaly detection (ELA, clone detection)
# numpy>=1.26.0              # Numerical processing for pixel analysis

# ── Report generation (optional) ──────────────────────────────────────────
# reportlab>=4.1.0           # PDF report generation
# weasyprint>=61.0           # HTML → PDF alternative

# ── Testing ────────────────────────────────────────────────────────────────
# httpx>=0.27.0              # async test client for FastAPI
# pytest>=8.1.0
# pytest-asyncio>=0.23.0
